% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{add_weights}
\alias{add_weights}
\title{Add weights to a dataset of responses}
\usage{
add_weights(
  data,
  params,
  weights = c("step1", "full"),
  add_weekly_weights = FALSE
)
}
\arguments{
\item{data}{a data frame containing a column called "token"}

\item{params}{a named list containing value "weights_in_dir", indicating where the
weights files are stored, and "weekly_weights_in_dir", indicating
where the weekly weights files are stored, if `add_weekly_weights` is TRUE}

\item{weights}{Which weights to use -- step1 or full?}

\item{add_weekly_weights}{boolean indicating whether to add weekly partial and full
weights in addition to daily weights}
}
\description{
There are two types of weights: step 1 weights and full weights. Step 1
weights are used for producing our aggregations; full weights are used
exclusively for producing individual response files.
}
