% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{apply_count_smoothing}
\alias{apply_count_smoothing}
\title{Applying smoothing to household counts data (CLI and ILI)}
\usage{
apply_count_smoothing(df, params, k = 3L, max_window = c(1L, 7L))
}
\arguments{
\item{df}{input data frame of summerized data}

\item{params}{a named list containing entried "start_time" and "end_time"}

\item{k}{integer}

\item{max_window}{pair of integers}
}
\description{
Applying smoothing to household counts data (CLI and ILI)
}
