% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_privacy.R
\name{apply_privacy_censoring}
\alias{apply_privacy_censoring}
\title{Censor aggregates to ensure privacy.}
\usage{
apply_privacy_censoring(df, params)
}
\arguments{
\item{df}{a data frame of summarized response data}

\item{params}{a named list with entries "s_weight", "s_mix_coef",
"num_filter"}
}
\description{
Currently done in simple, static way: Rows with sample size less than num_filter
are removed; no noise is added.
}
