% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert}
\alias{assert}
\title{Assert a logical value}
\usage{
assert(statement, msg = "")
}
\arguments{
\item{statement}{a logical value}

\item{msg}{a character string displayed as an additional message}
}
\description{
Will issue a \code{stop} command if the given statement is false.
}
