% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary_se}
\alias{binary_se}
\title{Calculate the standard error for a binary proportion (as a percentage)}
\usage{
binary_se(val, sample_size)
}
\arguments{
\item{val}{Vector of estimated percentages}

\item{sample_size}{Vector of corresponding sample sizes}
}
\value{
Vector of standard errors; NA when a sample size is 0.
}
\description{
Calculate the standard error for a binary proportion (as a percentage)
}
