% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bodge_B13}
\alias{bodge_B13}
\title{Fix B13 name in Wave 11.}
\usage{
bodge_B13(input_data, wave)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}

\item{wave}{integer indicating survey version}
}
\value{
corrected data frame
}
\description{
Fix B13 name in Wave 11.
}
