% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bodge_C6_C8}
\alias{bodge_C6_C8}
\title{Fix column names in Wave 10.}
\usage{
bodge_C6_C8(input_data)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}
}
\value{
corrected data frame
}
\description{
In Wave 10's deployment, the meaning of items C6 and C8 changed (from "In the
past 5 days, have you traveled outside of your state?" and "In the past 5
days, how often have you... felt depressed?", etc, to "In the past 7
days..."), but the names were not changed. The names are changed in later
waves.
}
\details{
We rename C6 and C8_\* to C6a and C8a_\*, respectively, to match the existing
naming scheme.
}
