% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bodge_E1}
\alias{bodge_E1}
\title{Fix E1_* names in Wave 11 data after ~June 16, 2021.}
\usage{
bodge_E1(input_data, wave)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}

\item{wave}{integer indicating survey version}
}
\value{
corrected data frame
}
\description{
Fix E1_* names in Wave 11 data after ~June 16, 2021.
}
