% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bodge_V2a}
\alias{bodge_V2a}
\title{Fix column names in Wave 13.}
\usage{
bodge_V2a(input_data, wave)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}

\item{wave}{integer indicating survey version}
}
\value{
corrected data frame
}
\description{
In Wave 13, a new item ("How many initial doses or shots did you receive of a
COVID-19 vaccine?") was added under the name V2a. However, a different item
("Did you receive (or do you plan to receive) all recommended doses?") was
asked under the name V2a in Waves 8-10. To differentiate, use the name V2d
for the newer item.
}
