% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bodge_v4_translation}
\alias{bodge_v4_translation}
\title{Fix translation error in Wave 6.}
\usage{
bodge_v4_translation(input_data)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}
}
\value{
corrected data frame, where V4 is the authoritative column
}
\description{
In Wave 6's first deployment, some of the translations swapped the order of
responses in V4, so these responses can't be interpreted correctly. Rather
than recoding them, we simply delete non-English translations of V4.
}
\details{
The updated deployment replaces V4 with V4a with correct translations. We
delete non-English V4 responses, then use V4a in place of V4 when present.
}
