% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{ceiling_epiweek}
\alias{ceiling_epiweek}
\title{Get date of the first day of the epiweek following the epiweek `x` falls in}
\usage{
ceiling_epiweek(date)
}
\arguments{
\item{date}{date}
}
\value{
date
}
\description{
Get date of the first day of the epiweek following the epiweek `x` falls in
}
