% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_addl_demographic}
\alias{code_addl_demographic}
\title{Demographics}
\usage{
code_addl_demographic(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Demographics
}
