% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_addl_symptoms}
\alias{code_addl_symptoms}
\title{COVID symptoms with modifications for contingency tables}
\usage{
code_addl_symptoms(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
COVID symptoms with modifications for contingency tables
}
