% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_age}
\alias{code_age}
\title{Age-related fields}
\usage{
code_age(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Age-related fields
}
