% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_attempt_vaccine}
\alias{code_attempt_vaccine}
\title{Attempts to get vaccinated}
\usage{
code_attempt_vaccine(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
data frame augmented with demographic grouping variables
}
\description{
Attempts to get vaccinated
}
