% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_children}
\alias{code_children}
\title{Misc children}
\usage{
code_children(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Misc children
}
