% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_education}
\alias{code_education}
\title{Education}
\usage{
code_education(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Education
}
