% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_gender}
\alias{code_gender}
\title{Gender}
\usage{
code_gender(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Gender
}
