% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_health}
\alias{code_health}
\title{Health conditions and behaviors}
\usage{
code_health(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Health conditions and behaviors
}
