% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_hh_size}
\alias{code_hh_size}
\title{Household total size}
\usage{
code_hh_size(input_data)
}
\arguments{
\item{input_data}{input data frame of raw survey data}
}
\value{
data frame augmented with `hh_number_total`
}
\description{
Household total size
}
