% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_mask_contact}
\alias{code_mask_contact}
\title{Mask and contact variables}
\usage{
code_mask_contact(input_data)
}
\arguments{
\item{input_data}{input data frame of raw survey data}
}
\value{
data frame augmented with `c_travel_state`, `c_work_outside_5d`,
  `c_mask_often`, `c_others_masked`
}
\description{
Mask and contact variables
}
