% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_mental_health}
\alias{code_mental_health}
\title{Mental health variables}
\usage{
code_mental_health(input_data)
}
\arguments{
\item{input_data}{input data frame of raw survey data}
}
\value{
data frame augmented with `mh_worried_ill`, `mh_anxious`,
  `mh_depressed`, `mh_isolated`, `mh_worried_finances`
}
\description{
Per our IRB, we only aggregate these variables for wave >= 4.
}
