% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_news_and_info}
\alias{code_news_and_info}
\title{COVID news and information variables}
\usage{
code_news_and_info(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
COVID news and information variables
}
