% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_numeric_freeresponse}
\alias{code_numeric_freeresponse}
\title{Convert a single free response column to numeric.}
\usage{
code_numeric_freeresponse(df, aggregations, col_var)
}
\arguments{
\item{df}{Data frame of individual response data.}

\item{aggregations}{Data frame with columns `name`, `var_weight`, `metric`,
`group_by`, `compute_fn`, `post_fn`. Each row represents one aggregate
to report. `name` is the aggregate's base column name; `var_weight` is the
column to use for its weights; `metric` is the column of `df` containing the
response value. `group_by` is a list of variables used to perform the
aggregations over. `compute_fn` is the function that computes
the aggregate response given many rows of data. `post_fn` is applied to the
aggregate data after megacounty aggregation, and can perform any final
calculations necessary.}

\item{col_var}{Name of response var}
}
\value{
list of data frame of individual response data and user-set data
frame of desired aggregations
}
\description{
Convert a single free response column to numeric.
}
