% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_race_ethnicity}
\alias{code_race_ethnicity}
\title{Race/ethnicity}
\usage{
code_race_ethnicity(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Race/ethnicity
}
