% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_schooling}
\alias{code_schooling}
\title{Schooling}
\usage{
code_schooling(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
data frame augmented with `hh_number_total`
}
\description{
Schooling
}
