% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_symptoms}
\alias{code_symptoms}
\title{Household symptom variables}
\usage{
code_symptoms(input_data)
}
\arguments{
\item{input_data}{input data frame of raw survey data}
}
\value{
data frame augmented with `hh_fever`, `hh_sore_throat`, `hh_cough`,
  `hh_short_breath`, `hh_diff_breath`, `hh_number_sick`
}
\description{
Household symptom variables
}
