% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_testing}
\alias{code_testing}
\title{Testing and test positivity variables}
\usage{
code_testing(input_data)
}
\arguments{
\item{input_data}{input data frame of raw survey data}
}
\value{
data frame augmented with `t_tested_14d`, `t_tested_positive_14d`,
  `t_wanted_test_14d`
}
\description{
Testing and test positivity variables
}
