% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_trust}
\alias{code_trust}
\title{Trust in various individuals and organizations for recommendations and info}
\usage{
code_trust(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Trust in various individuals and organizations for recommendations and info
}
