% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_vaccinated_breakdown}
\alias{code_vaccinated_breakdown}
\title{Vaccinated summary variable for making table cuts}
\usage{
code_vaccinated_breakdown(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Vaccinated summary variable for making table cuts
}
