% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{code_vaccine_barriers}
\alias{code_vaccine_barriers}
\title{Vaccination barriers}
\usage{
code_vaccine_barriers(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Vaccination barriers
}
