% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_vaccines}
\alias{code_vaccines}
\title{COVID vaccination variables}
\usage{
code_vaccines(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
data frame augmented with `v_covid_vaccinated` and
  `v_accept_covid_vaccine`
}
\description{
COVID vaccination variables
}
