% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R, R/contingency_tables.R
\name{compute_count}
\alias{compute_count}
\title{Returns multiple choice response estimates}
\usage{
compute_count(response, weight, sample_size)

compute_count(response, weight, sample_size)
}
\arguments{
\item{response}{a vector of multiple choice responses}

\item{weight}{a vector of sample weights for inverse probability weighting;
invariant up to a scaling factor}

\item{sample_size}{The sample size to use, which may be a non-integer (as
responses from ZIPs that span geographical boundaries are weighted
proportionately, and survey weights may also be applied)}
}
\value{
a vector of counts

a vector of counts
}
\description{
This function takes vectors as input and computes the response values
(a point estimate named "val" and a sample size
named "sample_size").

This function takes vectors as input and computes the response values
(a point estimate named "val" and a sample size
named "sample_size").
}
