% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_calculate.R
\name{compute_household_binary}
\alias{compute_household_binary}
\title{Wrapper for `compute_count_response` that adds sample_size. Val is the mean.}
\usage{
compute_household_binary(response, weight, sample_size, total_represented)
}
\arguments{
\item{response}{a vector of percentages (100 * cnt / total)}

\item{weight}{a vector of sample weights for inverse probability weighting;
invariant up to a scaling factor}

\item{sample_size}{The sample size to use, which may be a non-integer (as
responses from ZIPs that span geographical boundaries are weighted
proportionately, and survey weights may also be applied)}

\item{total_represented}{Number of people represented in sample, which may
be a non-integer}
}
\value{
a list of named means and other descriptive statistics
}
\description{
Wrapper for `compute_count_response` that adds sample_size. Val is the mean.
}
