% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{compute_mean}
\alias{compute_mean}
\title{Wrapper for `compute_count_response` that adds sample_size}
\usage{
compute_mean(response, weight, sample_size)
}
\arguments{
\item{response}{a vector of percentages (100 * cnt / total)}

\item{weight}{a vector of sample weights for inverse probability weighting;
invariant up to a scaling factor}

\item{sample_size}{Unused.}
}
\description{
Wrapper for `compute_count_response` that adds sample_size
}
