% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_calculate.R
\name{compute_numeric}
\alias{compute_numeric}
\title{Wrapper for `compute_count_response` that adds sample_size. Val is the mean.}
\usage{
compute_numeric(response, weight, sample_size)
}
\arguments{
\item{response}{a vector of percentages (100 * cnt / total)}

\item{weight}{a vector of sample weights for inverse probability weighting;
invariant up to a scaling factor}

\item{sample_size}{Unused.}
}
\value{
a vector of mean values
}
\description{
Wrapper for `compute_count_response` that adds sample_size. Val is the mean.
}
