% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_calculate.R
\name{compute_numeric_mean}
\alias{compute_numeric_mean}
\title{Return numeric response estimates. Val is the mean of a numeric vector.}
\usage{
compute_numeric_mean(response, weight, sample_size, total_represented)
}
\arguments{
\item{response}{a vector of multiple choice responses}

\item{weight}{a vector of sample weights for inverse probability weighting;
invariant up to a scaling factor}

\item{sample_size}{The sample size to use, which may be a non-integer (as
responses from ZIPs that span geographical boundaries are weighted
proportionately, and survey weights may also be applied)}

\item{total_represented}{Number of people represented in sample, which may
be a non-integer}
}
\value{
a list of named mean and other descriptive statistics
}
\description{
This function takes vectors as input and computes the response values
(a point estimate named "val" and 25, 50, and 75th percentiles).
}
