% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{compute_prop}
\alias{compute_prop}
\title{Wrapper for `compute_binary_response` that adds sample_size}
\usage{
compute_prop(response, weight, sample_size)
}
\arguments{
\item{response}{a vector of binary (0 or 1) responses}

\item{weight}{a vector of sample weights for inverse probability weighting;
invariant up to a scaling factor}

\item{sample_size}{The sample size to use, which may be a non-integer (as
responses from ZIPs that span geographical boundaries are weighted
proportionately, and survey weights may also be applied)}
}
\description{
Wrapper for `compute_binary_response` that adds sample_size
}
