% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables.R
\name{convert_3_binary_to_bool}
\alias{convert_3_binary_to_bool}
\title{Convert a single binary response + "I don't know" column to boolean}
\usage{
convert_3_binary_to_bool(df, col_var, yes_val = 1, no_val = 2, idk_val = 3)
}
\arguments{
\item{df}{Data frame of individual response data.}

\item{col_var}{Name of response var}

\item{yes_val}{Response code corresponding to answering "yes"}

\item{no_val}{Response code corresponding to answering "yes"}

\item{idk_val}{Response code corresponding to answering "I don't know"}
}
\value{
list of data frame of individual response data with newly mapped column
}
\description{
"I don't know" is set to NA (missing) and excluded.
}
