% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{convert_multiselect_to_binary_cols}
\alias{convert_multiselect_to_binary_cols}
\title{Convert a single multi-select response column to a set of boolean columns}
\usage{
convert_multiselect_to_binary_cols(df, aggregations, col_var)
}
\arguments{
\item{df}{Data frame of individual response data.}

\item{aggregations}{Data frame with columns `name`, `var_weight`, `metric`,
`compute_fn`, `post_fn`. Each row represents one aggregate
to report. `name` is the aggregate's base file name; `var_weight` is the column
to use for its weights; `metric` is the column of `df` containing the
response value. `compute_fn` is the function that computes
the aggregate response given many rows of data. `post_fn` is applied to the
aggregate data after megacounty aggregation, and can perform any final
calculations necessary.}

\item{col_var}{Name of response var}
}
\description{
Convert a single multi-select response column to a set of boolean columns
}
