% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_indicators.R
\name{create_aggs_product}
\alias{create_aggs_product}
\title{Create aggs from all combinations of provided input sets}
\usage{
create_aggs_product(regions, groups, indicators)
}
\arguments{
\item{regions}{list of strings indicating geo level}

\item{groups}{list of character vectors indicating grouping variables}

\item{indicators}{tibble with `name`, `metric`, `compute_fn`, and `post_fn`
columns}
}
\value{
tibble of created aggs
}
\description{
Create aggs from all combinations of provided input sets
}
