% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{create_complete_responses}
\alias{create_complete_responses}
\title{Create dataset for sharing with research partners}
\usage{
create_complete_responses(input_data, county_crosswalk)
}
\arguments{
\item{input_data}{data frame of responses}

\item{county_crosswalk}{crosswalk mapping ZIP5 to counties}
}
\description{
Different survey waves may have different sets of questions. Here we report
all questions across any wave, along with a wave identifier so analysts know
which wave a user took.
}
