% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{create_data_for_aggregatation}
\alias{create_data_for_aggregatation}
\title{Create variables needed for aggregation}
\usage{
create_data_for_aggregatation(input_data)
}
\arguments{
\item{input_data}{the input data frame of (filtered) responses}
}
\description{
Create variables needed for aggregation
}
