% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{create_derivative_columns}
\alias{create_derivative_columns}
\title{Create new columns for use in aggregates.}
\usage{
create_derivative_columns(df)
}
\arguments{
\item{df}{Data frame of individual response data.}
}
\value{
data frame of individual response data with newly derived columns
}
\description{
Create new columns for use in aggregates.
}
