% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{end_of_prev_full_month}
\alias{end_of_prev_full_month}
\title{Get the date of the last day of the previous month}
\usage{
end_of_prev_full_month(date)
}
\arguments{
\item{date}{Date that output will be calculated relative to}
}
\value{
Date
}
\description{
Get the date of the last day of the previous month
}
