% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{epiweek<-}
\alias{epiweek<-}
\title{epiweek equivalent of `week<-` as shown [here](https://lubridate.tidyverse.org/reference/week.html#examples)}
\usage{
epiweek(x) <- value
}
\arguments{
\item{x}{a date-time object. Must be a POSIXct, POSIXlt, Date, chron, 
yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or 
fts object.}

\item{value}{a numeric object}
}
\value{
date
}
\description{
epiweek equivalent of `week<-` as shown [here](https://lubridate.tidyverse.org/reference/week.html#examples)
}
