% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{experimental_arm_assignment}
\alias{experimental_arm_assignment}
\title{Label arms of experimental Wave 12.}
\usage{
experimental_arm_assignment(input_data, wave)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}

\item{wave}{integer indicating survey version}
}
\value{
data frame with new `module` column
}
\description{
Label arms of experimental Wave 12.
}
