% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{filter_data_for_aggregatation}
\alias{filter_data_for_aggregatation}
\title{Filter data that is appropriate for aggregation}
\usage{
filter_data_for_aggregatation(df)
}
\arguments{
\item{df}{data frame of responses}
}
\description{
Filter data that is appropriate for aggregation
}
