% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{filter_data_for_aggregation}
\alias{filter_data_for_aggregation}
\title{Filter data that is appropriate for aggregation}
\usage{
filter_data_for_aggregation(df, params, lead_days = 12L)
}
\arguments{
\item{df}{data frame of responses}

\item{params}{list containing `static_dir`, indicating where to find ZIP data
files, and `start_date`, indicating the first day for which estimates
should be produced}

\item{lead_days}{Integer specifying how many days of data *before*
`start_date` should be included in the data frame for aggregation. For
example, if we expect up to four days of survey backfill and seven days of
smoothing, we'd want to include at least 11 days of data before
`start_date`, so estimates on `start_date` are based on the correct data.}
}
\description{
Filter data that is appropriate for aggregation
}
