% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{filter_data_for_contingency}
\alias{filter_data_for_contingency}
\title{Filter out low-quality data for contingency tables}
\usage{
filter_data_for_contingency(df)
}
\arguments{
\item{df}{data frame of responses}
}
\description{
Filter out low-quality data for contingency tables
}
