% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{get_aggs}
\alias{get_aggs}
\title{Wrapper for `set_aggs`}
\usage{
get_aggs(params)
}
\arguments{
\item{params}{Named list of configuration parameters.}
}
\value{
a tibble of desired aggregations to calculate
}
\description{
Wrapper for `set_aggs`
}
