% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{get_binary_indicators}
\alias{get_binary_indicators}
\title{Fetch binary indicators to report in aggregates}
\usage{
get_binary_indicators()
}
\description{
Fetch binary indicators to report in aggregates
}
