% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_utils.R
\name{get_date_range_from_filenames}
\alias{get_date_range_from_filenames}
\title{Get date range based on list of input files provided.}
\usage{
get_date_range_from_filenames(params)
}
\arguments{
\item{params}{Params object produced by read_params}
}
\value{
Unnamed list of two dates
}
\description{
Get date range based on list of input files provided.
}
