% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{get_hh_count_indicators}
\alias{get_hh_count_indicators}
\title{Fetch data frame specifying all count indicators to report aggregates.}
\usage{
get_hh_count_indicators()
}
\description{
Fetch data frame specifying all count indicators to report aggregates.
}
