% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{get_range_prev_full_month}
\alias{get_range_prev_full_month}
\title{Get the date range specifying the previous month}
\usage{
get_range_prev_full_month(date = Sys.Date())
}
\arguments{
\item{date}{Date that output will be calculated relative to}
}
\value{
list of two Dates
}
\description{
Get the date range specifying the previous month
}
