% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{get_range_prev_full_period}
\alias{get_range_prev_full_period}
\title{Get the date range specifying the previous full time period.}
\usage{
get_range_prev_full_period(date = Sys.Date(), weekly_or_monthly_flag)
}
\arguments{
\item{date}{Date of interest}

\item{weekly_or_monthly_flag}{string "week" or "month" indicating desired
time period to aggregate over}
}
\value{
list of two Dates
}
\description{
Get the date range specifying the previous full time period.
}
