% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{get_range_prev_full_week}
\alias{get_range_prev_full_week}
\title{Get the date range specifying the previous week}
\usage{
get_range_prev_full_week(date = Sys.Date())
}
\arguments{
\item{date}{Date of interest}
}
\value{
list of two Dates
}
\description{
Get the date range specifying the previous week
}
