% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_utils.R
\name{get_sparse_filenames}
\alias{get_sparse_filenames}
\title{Get sparse list of input data files from `input_dir`.}
\usage{
get_sparse_filenames(start_date, end_date, params)
}
\arguments{
\item{start_date}{Start of desired date range}

\item{end_date}{End of desired date range}

\item{params}{Params object produced by read_params}
}
\value{
Character vector of filenames
}
\description{
Finds every fourth + last file by date.
}
