% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{jeffreys_binary}
\alias{jeffreys_binary}
\title{Apply a Jeffreys correction to estimates and their standard errors.}
\usage{
jeffreys_binary(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Updated data frame.
}
\description{
Apply a Jeffreys correction to estimates and their standard errors.
}
