% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{jeffreys_count}
\alias{jeffreys_count}
\title{Apply a Jeffreys correction to standard errors.}
\usage{
jeffreys_count(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Updated data frame.
}
\description{
See `jeffreys_se()` for reasoning and methods.
}
