% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{jeffreys_multinomial_factory}
\alias{jeffreys_multinomial_factory}
\title{Generate function that applies Jeffreys correction to multinomial estimates.}
\usage{
jeffreys_multinomial_factory(k)
}
\arguments{
\item{k}{Number of groups.}
}
\value{
Function to apply multinomial Jeffreys correction.
}
\description{
Generate function that applies Jeffreys correction to multinomial estimates.
}
