% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{jeffreys_multinomial_percentage}
\alias{jeffreys_multinomial_percentage}
\title{Adjust a multionmial percentage estimate to use the Jeffreys method.}
\usage{
jeffreys_multinomial_percentage(percentage, sample_size)
}
\arguments{
\item{percentage}{Vector of percentages to adjust.}

\item{sample_size}{Vector of corresponding sample sizes.}
}
\value{
Vector of adjusted percentages.
}
\description{
Takes a previously estimated percentage (calculated with num_group1 / total *
100) and replaces it with the Jeffreys version, where one pseudo-observation
with 1/k * 100 % group1 is inserted.
}
