% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{join_weights}
\alias{join_weights}
\title{Add weights to a dataset of responses}
\usage{
join_weights(data, params)
}
\arguments{
\item{data}{a data frame containing a column called "token"}

\item{params}{a named list containing a value "weights_in_dir" indicating where the
weights files are stored}
}
\description{
Add weights to a dataset of responses
}
