% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{load_and_join_weights}
\alias{load_and_join_weights}
\title{Add a single type of weights to a dataset of responses as field `weight`}
\usage{
load_and_join_weights(data, weights_dir, pattern)
}
\arguments{
\item{data}{a data frame containing a column called "token"}

\item{weights_dir}{directory to look for the weights in}

\item{pattern}{regular expression matching desired weights files}
}
\description{
Add a single type of weights to a dataset of responses as field `weight`
}
