% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{load_responses_all}
\alias{load_responses_all}
\title{Load all response datasets in a local directory}
\usage{
load_responses_all(params)
}
\arguments{
\item{params}{a named listed containing a value named "input", a vector of
paths to load by the function, and "input_dir", the directory where the
input files are found}
}
\value{
A data frame of all loaded data files concatenated into one data
  frame
}
\description{
Note that if some columns are not present in all files -- for example, if
survey questions changed and so newer data files have different columns --
the resulting data frame will contain all columns, with NAs in rows where
that column was not present.
}
