% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_megacounty_fips}
\alias{make_megacounty_fips}
\title{Converts county FIPS code to megacounty code.}
\usage{
make_megacounty_fips(fips)
}
\arguments{
\item{fips}{Geo-id}
}
\value{
Megacounty
}
\description{
We designate megacounties with a special FIPS ending in 000; for example, the
megacounty for state 26 would be 26000 and would comprise counties with FIPS
codes 26XXX.
}
