% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{merge_responses}
\alias{merge_responses}
\title{Merge new data with archived data}
\usage{
merge_responses(input_data, archive)
}
\arguments{
\item{input_data}{data frame containing the new response data}

\item{archive}{archive data read by `load_archive()`}
}
\value{
single data frame containing the merged data to be used for analysis
}
\description{
See the README for details on how the archive works and why it is necessary
to use the archive to check submitted tokens.
}
