% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mix_weights}
\alias{mix_weights}
\title{Adjust weights so no weight is not too much of the final estimate}
\usage{
mix_weights(weights, params)
}
\arguments{
\item{weights}{a vector of sample weights}

\item{params}{a named list containing an element named "num_filter"; the maximum
weight is assumed to be 1 / "num_filter" * 0.999.}
}
\description{
Adjust weights so no weight is not too much of the final estimate
}
