% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{module_assignment}
\alias{module_assignment}
\title{Process module assignment column.}
\usage{
module_assignment(input_data, wave)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}

\item{wave}{integer indicating survey version}
}
\value{
data frame with new `module` column
}
\description{
Rename `module` and recode to A/B/`NA`. Note: module assignment column name
may change with survey version.
}
