% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_df}
\alias{msg_df}
\title{Write a message to the console}
\usage{
msg_df(text, df)
}
\arguments{
\item{text}{the body of the message to display}

\item{df}{a data frame; the message will show the number of rows in the data frame}
}
\description{
Write a message to the console
}
