% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{past_n_days}
\alias{past_n_days}
\title{Return vector from the past n days, inclusive}
\usage{
past_n_days(date, ndays = 0L)
}
\arguments{
\item{date}{a string containing a single date that can be parsed with `ymd`, such as
"20201215"}

\item{ndays}{how many days in the past to include}
}
\description{
Returns dates as strings in the form "YYYYMMDD"
}
