% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_calculate.R
\name{post_convert_count_to_pct}
\alias{post_convert_count_to_pct}
\title{Convert val column from counts to percents of total}
\usage{
post_convert_count_to_pct(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Updated data frame.
}
\description{
Meant for use as `post_fn` with multiple choice metrics, which produce
weighted frequency for val by default.
}
