% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_allowed_zip5}
\alias{produce_allowed_zip5}
\title{Returns list of zip codes with large enough population to include in output}
\usage{
produce_allowed_zip5(static_dir)
}
\arguments{
\item{static_dir}{local directory containing the file "02_20_uszips.csv"}
}
\description{
For privacy reasons, we only include zip codes with a population over 100 people.
}
