% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_crosswalk_county}
\alias{produce_crosswalk_county}
\title{Produce crosswalk file relating zip codes to counties}
\usage{
produce_crosswalk_county(static_dir, zip_metadata)
}
\arguments{
\item{static_dir}{local directory containing the file "02_20_uszips.csv"}

\item{zip_metadata}{output from a call to the function produce_allowed_zip5}
}
\value{
A tibble containing three columns: zip5, geo_id (county fips codes),
    and weight_in_location. Since ZIP codes can span multiple counties, the
    weight_in_location indicates the fraction of the ZIP population contained
    in the county. When grouped by ZIP5, this should sum to 1.
}
\description{
Produce crosswalk file relating zip codes to counties
}
