% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_crosswalk_hrr}
\alias{produce_crosswalk_hrr}
\title{Produce crosswalk file relating zip codes to hospital referral regions (HRRs)}
\usage{
produce_crosswalk_hrr(zip_metadata)
}
\arguments{
\item{zip_metadata}{output from a call to the function produce_allowed_zip5}
}
\value{
A tibble containing three columns: zip5, geo_id (HRR code as character), and
         weight_in_location. The variable weight_in_location will sum to 1 when grouped
         by zip5.
}
\description{
Produce crosswalk file relating zip codes to hospital referral regions (HRRs)
}
