% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_crosswalk_list}
\alias{produce_crosswalk_list}
\title{Produce a list of crosswalk files relating zip5s to geographic regions}
\usage{
produce_crosswalk_list(static_dir)
}
\arguments{
\item{static_dir}{local directory containing the file "02_20_uszips.csv"}
}
\value{
A list containing elements for four geographic regions: county, state, msa, and
         hrr (hospital referral region). Each element contains a tibble with three
         columns named "zip5", "geo_id", and "weight_in_location"
}
\description{
Produce a list of crosswalk files relating zip5s to geographic regions
}
