% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_crosswalk_msa}
\alias{produce_crosswalk_msa}
\title{Produce crosswalk file relating zip codes to metropolitan statistical areas (MSAs)}
\usage{
produce_crosswalk_msa(static_dir, crosswalk_county)
}
\arguments{
\item{static_dir}{local directory containing the file "02_20_uszips.csv"}

\item{crosswalk_county}{output from a call to the function produce_crosswalk_county}
}
\value{
A tibble containing three columns: zip5, geo_id (MSA fips codes), and
         weight_in_location. The variable weight_in_location will sum to 1 when grouped
         by zip5.
}
\description{
Produce crosswalk file relating zip codes to metropolitan statistical areas (MSAs)
}
