% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_crosswalk_national}
\alias{produce_crosswalk_national}
\title{Produce crosswalk file relating zip codes to the entire U.S.}
\usage{
produce_crosswalk_national(zip_metadata)
}
\arguments{
\item{zip_metadata}{output from a call to the function produce_allowed_zip5}
}
\value{
A tibble containing three columns: zip5, geo_id (county fips codes), and
         weight_in_location.
}
\description{
Every zip code is mapped to one entity, "usa".
}
