% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_crosswalk_state}
\alias{produce_crosswalk_state}
\title{Produce crosswalk file relating zip codes to states}
\usage{
produce_crosswalk_state(zip_metadata, crosswalk_county)
}
\arguments{
\item{zip_metadata}{output from a call to the function produce_allowed_zip5}

\item{crosswalk_county}{output from a call to the function produce_crosswalk_county}
}
\value{
A tibble containing three columns: zip5, geo_id (state two-letter codes), and
         weight_in_location. The variable weight_in_location will sum to 1 when grouped
         by zip5.
}
\description{
Produce crosswalk file relating zip codes to states
}
