% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{produce_zip_metadata}
\alias{produce_zip_metadata}
\title{Returns metadata about each U.S. Zip Code}
\usage{
produce_zip_metadata(static_dir)
}
\arguments{
\item{static_dir}{local directory containing the file "02_20_uszips.csv"}
}
\description{
Returns metadata about each U.S. Zip Code
}
