% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_utils.R
\name{read_contingency_params}
\alias{read_contingency_params}
\title{Return params file as an R list}
\usage{
read_contingency_params(
  path = "params.json",
  template_path = "params.json.template"
)
}
\arguments{
\item{path}{path to the parameters file; if not present, will try to copy the file
"params.json.template"}

\item{template_path}{path to the template parameters file}
}
\value{
a named list of parameters values
}
\description{
Reads a parameters file. Copies global params to contingency params if not
already defined.
}
