% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_params}
\alias{read_params}
\title{Return params file as an R list}
\usage{
read_params()
}
\value{
a named list of parameters values
}
\description{
Reads the file "params.json" in the current working directory. If the file does not
exist, the function will create a copy of "params.json.template" and read from that.
}
