% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{reformat_responses}
\alias{reformat_responses}
\title{Remap race and binary columns to make more interpretable}
\usage{
reformat_responses(df)
}
\arguments{
\item{df}{Data frame of individual response data.}
}
\value{
data frame of individual response data with newly mapped columns
}
\description{
Remap race and binary columns to make more interpretable
}
