% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{remap_responses}
\alias{remap_responses}
\title{Remap binary columns, race, and others to make more interpretable.}
\usage{
remap_responses(df)
}
\arguments{
\item{df}{Data frame of individual response data.}
}
\value{
data frame of individual response data with newly mapped columns
}
\description{
Remap binary columns, race, and others to make more interpretable.
}
