% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_variables.R
\name{rename_responses}
\alias{rename_responses}
\title{Rename all columns to make more interpretable.}
\usage{
rename_responses(df)
}
\arguments{
\item{df}{Data frame of individual response data.}
}
\value{
data frame of individual response data with newly mapped columns
}
\description{
Rename all columns to make more interpretable.
}
