% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_privacy.R
\name{round_n}
\alias{round_n}
\title{Round sample sizes to nearest 5.}
\usage{
round_n(df, params)
}
\arguments{
\item{df}{a data frame of summarized response data}

\item{params}{a named list with entries "s_weight", "s_mix_coef",
"num_filter"}
}
\description{
Round sample sizes to nearest 5.
}
