% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{run_contingency_tables}
\alias{run_contingency_tables}
\title{Run the contingency table production pipeline}
\usage{
run_contingency_tables(params, user_aggs)
}
\arguments{
\item{params}{Params object produced by read_params}

\item{user_aggs}{Tibble of user-specified aggregations to perform}
}
\value{
none
}
\description{
See the README.md file in the source directory for more information about how to run
this function.
}
