% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_facebook}
\alias{run_facebook}
\title{Run the entire pipeline}
\usage{
run_facebook()
}
\value{
none
}
\description{
See the README.md file in the source directory for more information about how to run
this function.
}
