% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{set_aggs}
\alias{set_aggs}
\title{Sets user-specified aggregations.}
\usage{
set_aggs(params)
}
\arguments{
\item{params}{Named list of configuration parameters.}
}
\value{
a tibble of desired aggregations to calculate
}
\description{
User should add additional desired aggregations here following existing 
format. Names should be unique. Listing no groupby vars will implicitly
compute aggregations at the national level.
}
\details{
Compute functions must be one of the `compute_*` set (or another function 
with similar format can be created).
}
