% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R, R/contingency_tables.R
\name{set_human_readable_colnames}
\alias{set_human_readable_colnames}
\title{Rename question codes to informative descriptions}
\usage{
set_human_readable_colnames(input_data)

set_human_readable_colnames(input_data)
}
\arguments{
\item{input_data}{Data frame of individual response data}
}
\value{
Data frame with descriptive column names

Data frame with descriptive column names
}
\description{
Column names beginning with "b_" are binary (T/F/NA); with "t_" are user-
entered text; with "n_" are user-entered numeric; with "mc_" are multiple
choice (where only a single response can be selected); and with "ms_" are
so-called multi-select, where multiple responses can be selected.

Column names beginning with "b_" are binary (T/F/NA); with "t_" are user-
entered text; with "n_" are user-entered numeric; with "mc_" are multiple
choice (where only a single response can be selected); and with "ms_" are
so-called multi-select, where multiple responses can be selected.
}
\details{
Only binary columns are mapped from response codes to real values. Multiple
choice and multi-select questions use the original numeric response codes.

Only binary columns are mapped from response codes to real values. Multiple
choice and multi-select questions use the original numeric response codes.
}
