% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{split_options}
\alias{split_options}
\title{Split multiselect options into codable form}
\usage{
split_options(column)
}
\arguments{
\item{column}{vector of selections, like c("1,4", "5", ...)}
}
\value{
list of same length, each entry of which is a vector of selected
  options
}
\description{
Multiselect options are coded by Qualtrics as a comma-separated string of
selected options, like "1,14", or the empty string if no options are
selected. Split these into vectors of selected options.
}
