% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables.R
\name{start_of_month}
\alias{start_of_month}
\title{Get the date of the first of the previous month}
\usage{
start_of_month(date)
}
\arguments{
\item{end_date}{a}
}
\description{
Get the date of the first of the previous month
}
