% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{start_of_prev_full_week}
\alias{start_of_prev_full_week}
\title{Get the date of the first day of the previous week}
\usage{
start_of_prev_full_week(date)
}
\arguments{
\item{date}{Date that output will be calculated relative to}
}
\value{
Date
}
\description{
Get the date of the first day of the previous week
}
