% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{sum_n_days}
\alias{sum_n_days}
\title{Smooth data by summing responses n days in the past}
\usage{
sum_n_days(df, smooth_days, params)
}
\arguments{
\item{df}{a data frame of survey responses}

\item{smooth_days}{integer; how many days in the past to smooth ?}

\item{params}{a named list with entries "start_time", and "end_time"}
}
\description{
Smooth data by summing responses n days in the past
}
