% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{summarize_aggs}
\alias{summarize_aggs}
\title{Calculate aggregates across all days for all indicators.}
\usage{
summarize_aggs(df, crosswalk_data, aggregations, geo_level, params)
}
\arguments{
\item{df}{a data frame of survey responses}

\item{crosswalk_data}{An aggregation, such as zip => county or zip => state,
as a data frame with a "zip5" column to join against.}

\item{aggregations}{Data frame of desired aggregations.}

\item{geo_level}{the aggregation level, such as county or state, being used}

\item{params}{a named list with entries "s_weight", "s_mix_coef",
"num_filter"}
}
\description{
The organization may seem a bit contorted, but this is designed for speed.
The primary bottleneck is repeatedly filtering the data frame to find data
for the grouping variables of interest. To save time, we do this once
and then calculate all indicators for that groupby combination, rather than
separately filtering every time we want to calculate a new table. We also
rely upon data.table's keys and indices to allow us to do the filtering in
O(log n) time, which is important when the data frame contains millions of
rows.
}
