% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{summarize_binary_day}
\alias{summarize_binary_day}
\title{Produce an estimate for all binary indicators on a specific target day.}
\usage{
summarize_binary_day(
  day_df,
  crosswalk_data,
  indicators,
  target_day,
  geo_level,
  params
)
}
\arguments{
\item{day_df}{Data frame containing all data needed to estimate one day.
Estimates for `target_day` will be based on all of this data, so if the
estimates are meant to be moving averages, `day_df` may contain multiple
days of data.}

\item{crosswalk_data}{See `summarize_binary()`. This contains the crosswalk
for just one `geo_level`, rather than all crosswalks.}

\item{indicators}{See `summarize_binary()`.}

\item{target_day}{A `Date` indicating the day for which these estimates are
to be calculated.}

\item{geo_level}{Name of the geo level (county, state, etc.) for which we are
aggregating.}

\item{params}{Named list of configuration options.}
}
\description{
Produce an estimate for all binary indicators on a specific target day.
}
