% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{summarize_hh_count}
\alias{summarize_hh_count}
\title{Summarize CLI and ILI household variables at a geographic level}
\usage{
summarize_hh_count(
  df,
  crosswalk_data,
  metric,
  var_weight,
  params,
  smooth = FALSE
)
}
\arguments{
\item{df}{a data frame of survey responses}

\item{crosswalk_data}{a named list containing geometry crosswalk files from zip5 values}

\item{metric}{name of the metric to use; should be "ili" or "cli"}

\item{var_weight}{name of the variable containing the survey weights}

\item{params}{a named list with entries "s_weight", "s_mix_coef", "num_filter",
"start_time", and "end_time"}

\item{smooth}{logical; should the results by smoothed?}
}
\description{
Summarize CLI and ILI household variables at a geographic level
}
