% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{summarize_indicators}
\alias{summarize_indicators}
\title{Calculate aggregates across all days for all indicators.}
\usage{
summarize_indicators(df, crosswalk_data, indicators, geo_level, days, params)
}
\arguments{
\item{df}{a data frame of survey responses}

\item{crosswalk_data}{An aggregation, such as zip => county or zip => state,
as a data frame with a "zip5" column to join against.}

\item{indicators}{Data frame of indicators.}

\item{geo_level}{the aggregation level, such as county or state, being used}

\item{days}{a vector of Dates for which we should generate response estimates}

\item{params}{a named list with entries "s_weight", "s_mix_coef",
"num_filter"}
}
\description{
The organization may seem a bit contorted, but this is designed for speed.
The primary bottleneck is repeatedly filtering the data frame to find data
for the day and geographic area of interest. To save time, we do this once
and then calculate all indicators for that day-area combination, rather than
separately filtering every time we want to calculate a new indicator. We also
rely upon data.table's keys and indices to allow us to do the filtering in
O(log n) time, which is important when the data frame contains millions of
rows.
}
