% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{summarize_indicators_day}
\alias{summarize_indicators_day}
\title{Produce estimates for all indicators on a specific target day.}
\usage{
summarize_indicators_day(day_df, indicators, target_day, geo_level, params)
}
\arguments{
\item{day_df}{Data frame containing all data needed to estimate one day.
Estimates for `target_day` will be based on all of this data, so if the
estimates are meant to be moving averages, `day_df` may contain multiple
days of data.}

\item{indicators}{Indicators to report. See `aggregate_indicators()`.}

\item{target_day}{A `Date` indicating the day for which these estimates are
to be calculated.}

\item{geo_level}{Name of the geo level (county, state, etc.) for which we are
aggregating.}

\item{params}{Named list of configuration options.}
}
\description{
Produce estimates for all indicators on a specific target day.
}
