% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{surveyID_to_wave}
\alias{surveyID_to_wave}
\title{Map Qualtrics survey IDs to wave number.}
\usage{
surveyID_to_wave(surveyID)
}
\arguments{
\item{surveyID}{Qualtrics Survey ID}
}
\value{
Wave number (integer), or NA if the survey ID is unknown
}
\description{
Wave numbers are documented on our survey documentation site with full coding
details, so we can include wave number in the individual output files so
users know which wave the user completed.
}
