% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R, R/contingency_tables.R
\name{update_params}
\alias{update_params}
\title{Update date and input files settings from params file}
\usage{
update_params(params)

update_params(params)
}
\arguments{
\item{params}{Params object produced by read_params}
}
\value{
a named list of parameters values

a named list of parameters values
}
\description{
Use `end_date` and `aggregate_range` to calculate last full time period
(either month or week) to use for aggregations. Find all files in `input_dir`
within the calculated time period.

Use `end_date` and `aggregate_range` to calculate last full time period
(either month or week) to use for aggregations. Find all files in `input_dir`
within the calculated time period.
}
