% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_utils.R
\name{verify_aggs}
\alias{verify_aggs}
\title{Checks user-set aggregations for basic validity}
\usage{
verify_aggs(aggregations)
}
\arguments{
\item{aggregations}{Data frame with columns `name`, `var_weight`, `metric`,
`group_by`, `compute_fn`, `post_fn`. Each row represents one aggregate
to report. `name` is the aggregate's base column name; `var_weight` is the 
column to use for its weights; `metric` is the column of `df` containing the
response value. `group_by` is a list of variables used to perform the 
aggregations over. `compute_fn` is the function that computes
the aggregate response given many rows of data. `post_fn` is applied to the
aggregate data after megacounty aggregation, and can perform any final
calculations necessary.}
}
\value{
a data frame of desired aggregations to calculate
}
\description{
Checks user-set aggregations for basic validity
}
