% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{write_binary_variables}
\alias{write_binary_variables}
\title{Write binary response variables for export to the API}
\usage{
write_binary_variables(df, cw_list, params)
}
\arguments{
\item{df}{a data frame of survey responses}

\item{cw_list}{a named list containing geometry crosswalk files from zip5 values}

\item{params}{a named list with entries "start_time", and "end_time"}
}
\description{
Write binary response variables for export to the API
}
