% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{write_cid}
\alias{write_cid}
\title{Write a file containing the new tokens seen in the dataset}
\usage{
write_cid(data, type_name, params, module_type = "", experimental_cids = FALSE)
}
\arguments{
\item{data}{a data frame containing a column named "token"}

\item{type_name}{character value used for naming the output file}

\item{params}{a named list; must contain entries "start_time", "end_time",
and "weights_out_dir" or "experimental_weights_out_dir". These are used in
constructing the path where the output data will be stored.}

\item{module_type}{character value used to indicate module filtering, if any}

\item{experimental_cids}{boolean flag indicating if CIDs to save should use
the, as of now, experimental format}
}
\description{
Write only tokens that appeared between `start_date` and `end_date`.
}
