% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{write_cid_experimental_wrapper}
\alias{write_cid_experimental_wrapper}
\title{Write a file containing newly-seen tokens, using new CID list names}
\usage{
write_cid_experimental_wrapper(data, type_name, params, module_type)
}
\arguments{
\item{data}{a data frame containing a column named "token"}

\item{type_name}{character value used for naming the output file}

\item{params}{a named list; must contain entries "start_time", "end_time",
and "weights_out_dir". These are used in constructing the path where the
output data will be stored.}

\item{module_type}{character value used to indicate module filtering, if any}
}
\description{
Write a file containing newly-seen tokens, using new CID list names
}
