% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_api.R
\name{write_data_api}
\alias{write_data_api}
\title{Write csv file for injestion into the Delphi API}
\usage{
write_data_api(data, params, geo_name, signal_name)
}
\arguments{
\item{data}{a data frame to save; must contain the columns "geo_id", "val",
"se", "sample_size", and "day". The first four are saved in the
output; day is used for spliting the data into files.}

\item{params}{a named list, containing the value "export_dir" indicating the
directory where the csv should be saved}

\item{geo_name}{name of the geographic level; used for naming the output file}

\item{signal_name}{name of the signal; used for naming the output file}
}
\description{
Write csv file for injestion into the Delphi API
}
