% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{write_hh_count_data}
\alias{write_hh_count_data}
\title{Write CLI and ILI counts data for export to the API}
\usage{
write_hh_count_data(df, cw_list, params)
}
\arguments{
\item{df}{a data frame of survey responses}

\item{cw_list}{a named list containing geometry crosswalk files from zip5 values}

\item{params}{a named list with entires "s_weight", "s_mix_coef", "num_filter",
"start_time", and "end_time"}
}
\description{
Write CLI and ILI counts data for export to the API
}
