% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{write_individual}
\alias{write_individual}
\title{Save individual response data}
\usage{
write_individual(data_full_w, params)
}
\arguments{
\item{data_full_w}{data frame of response values}

\item{params}{a named list containing an element named "individual_dir"
indicating where the output file will be written, and values
"start_time" and "end_time" used in the naming the output file}
}
\description{
The "token" column is removed from the input dataset, which is otherwise saved as-is.
}
