% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{write_individual_day}
\alias{write_individual_day}
\title{Save individual response data for a single date}
\usage{
write_individual_day(
  data_to_write,
  params,
  date,
  out_dir,
  raceeth_version = ""
)
}
\arguments{
\item{data_to_write}{data frame of response values}

\item{params}{a named list containing values "start_time"
and "end_time" used in naming the output file}

\item{date}{selected day to save data for}

\item{out_dir}{location where the output file will be written}
}
\description{
Save individual response data for a single date
}
