% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bodge_E1}
\alias{bodge_E1}
\title{Fix E1_* names in Wave 11 data after ~June 16, 2021.}
\usage{
bodge_E1(input_data, wave)
}
\arguments{
\item{input_data}{data frame of responses, before subsetting to select
variables}

\item{wave}{integer indicating survey version}
}
\value{
corrected data frame
}
\description{
Items E1_1 through E1_4 are part of a matrix. Qualtrics, for unknown reasons,
switched to naming these E1_4 through E1_7 in June. Convert back to the
intended names.
}
