% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_beliefs}
\alias{code_beliefs}
\title{Beliefs}
\usage{
code_beliefs(input_data)
}
\arguments{
\item{input_data}{input data frame of raw survey data}
}
\value{
augmented data frame
}
\description{
Beliefs
}
