% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code_mask_contact}
\alias{code_mask_contact}
\title{Mask and contact variables}
\usage{
code_mask_contact(input_data, wave)
}
\arguments{
\item{input_data}{input data frame of raw survey data}

\item{wave}{integer indicating survey version}
}
\value{
augmented data frame
}
\description{
Mask and contact variables
}
