% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{compute_binary_response}
\alias{compute_binary_response}
\title{Returns binary response estimates}
\usage{
compute_binary_response(response, weight, sample_weight)
}
\arguments{
\item{response}{a vector of binary (0 or 1) responses}

\item{weight}{a vector of sample weights for inverse probability
weighting; invariant up to a scaling factor}

\item{sample_weight}{a vector of sample size weights; the total sample size
will be the sum of these values}
}
\description{
This function takes vectors as input and computes the binary response values (a point
estimate named "val", a standard error named "se", and a sample size named "sample_size")
Note that there are two different sets of weights that have a different effect on the
output.
}
