% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{compute_count_response}
\alias{compute_count_response}
\title{Returns response estimates for a single geographic area.}
\usage{
compute_count_response(response, weight)
}
\arguments{
\item{response}{a vector of percentages (100 * cnt / total)}

\item{weight}{a vector of sample weights for inverse probability weighting;
invariant up to a scaling factor}
}
\description{
This function takes vectors as input and computes the count response values
(a point estimate named "val", a standard error named "se", and an effective
sample size named "effective_sample_size").
}
