% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{create_complete_responses}
\alias{create_complete_responses}
\title{Create dataset for sharing with research parteners}
\usage{
create_complete_responses(input_data)
}
\arguments{
\item{input_data}{data frame of responses}
}
\description{
Create dataset for sharing with research parteners
}
