% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{epiweek<-}
\alias{epiweek<-}
\title{epiweek equivalent of `week<-` as shown [here](https://lubridate.tidyverse.org/reference/week.html#examples)}
\usage{
epiweek(x) <- value
}
\description{
epiweek equivalent of `week<-` as shown [here](https://lubridate.tidyverse.org/reference/week.html#examples)
}
