% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{floor_epiweek}
\alias{floor_epiweek}
\title{Get date of the first day of the epiweek `x` falls in.}
\usage{
floor_epiweek(date)
}
\arguments{
\item{date}{date}
}
\value{
date
}
\description{
Get date of the first day of the epiweek `x` falls in.
}
