% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{get_aggs}
\alias{get_aggs}
\title{Checks user-set aggregations for basic validity}
\usage{
get_aggs(params, aggs)
}
\arguments{
\item{params}{Named list of configuration parameters.}
}
\value{
a tibble of desired aggregations to calculate
}
\description{
Checks user-set aggregations for basic validity
}
