% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_utils.R
\name{get_filenames_in_range}
\alias{get_filenames_in_range}
\title{Get relevant input data file names from `input_dir`.}
\usage{
get_filenames_in_range(start_date, end_date, params)
}
\arguments{
\item{start_date}{Start of desired date range}

\item{end_date}{End of desired date range}

\item{params}{Params object produced by read_params}
}
\value{
Character vector of filenames
}
\description{
Get relevant input data file names from `input_dir`.
}
