% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tables.R
\name{get_filenames_in_range}
\alias{get_filenames_in_range}
\title{Get relevant input data file names from `input_dir`.}
\usage{
get_filenames_in_range(date_range, params)
}
\arguments{
\item{date_range}{List of two dates specifying start and end of desired
date range}

\item{params}{Params object produced by read_params}
}
\value{
Character vector of filenames
}
\description{
Get relevant input data file names from `input_dir`.
}
