% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{is_selected}
\alias{is_selected}
\title{Test if a specific selection is selected}
\usage{
is_selected(vec, selection)
}
\arguments{
\item{vec}{A list whose entries are character vectors, such as c("14", "15").}

\item{selection}{one string, such as "14"}
}
\value{
a logical vector; for each list entry, whether selection is contained
  in the character vector.
}
\description{
Checking whether a specific selection is selected in either "" (empty
string) or `NA` responses will produce `NA`s.
}
