% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{jeffreys_se}
\alias{jeffreys_se}
\title{Apply Jeffrey's Prior to correct standard error values}
\usage{
jeffreys_se(old_se, percent, sample_size)
}
\arguments{
\item{old_se}{a numeric vector of previous standard errors}

\item{percent}{a numeric vector of the the estimated point estimates}

\item{sample_size}{a numeric vector of the sample sizes}
}
\description{
Apply Jeffrey's Prior to correct standard error values
}
