% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{load_archive}
\alias{load_archive}
\title{Load data from cached archive}
\usage{
load_archive(params)
}
\arguments{
\item{params}{a named list, read from read_params. Must contain and element named
archive_dir (containing the file "data_agg.Rds") and a logical value
named debug.}
}
\description{
Load data from cached archive
}
