% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{load_response_one}
\alias{load_response_one}
\title{Load a single set of responses}
\usage{
load_response_one(input_filename, params)
}
\arguments{
\item{input_filename}{filename of the input CSV file}

\item{params}{a named list containing a value named "input_dir", the directory
where the input file are found}
}
\description{
Load a single set of responses
}
