% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_utils.R
\name{msg_plain}
\alias{msg_plain}
\title{Write a message to the console}
\usage{
msg_plain(text)
}
\arguments{
\item{text}{the body of the message to display}
}
\description{
Write a message to the console
}
