% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R, R/contingency_tables.R
\name{msg_plain}
\alias{msg_plain}
\title{Write a time-stamped message to the console}
\usage{
msg_plain(text)

msg_plain(text)
}
\arguments{
\item{text}{the body of the message to display}
}
\description{
Write a time-stamped message to the console

Write a time-stamped message to the console
}
