% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R, R/contingency_tables.R
\name{post_process_aggs}
\alias{post_process_aggs}
\title{Post-process aggregations and data to make more generic}
\usage{
post_process_aggs(df, aggregations, cw_list)

post_process_aggs(df, aggregations, cw_list)
}
\arguments{
\item{df}{Data frame of individual response data.}

\item{aggregations}{Data frame with columns `name`, `var_weight`, `metric`,
`group_by`, `compute_fn`, `post_fn`. Each row represents one aggregate
to report. `name` is the aggregate's base column name; `var_weight` is the 
column to use for its weights; `metric` is the column of `df` containing the
response value. `group_by` is a list of variables used to perform the 
aggregations over. `compute_fn` is the function that computes
the aggregate response given many rows of data. `post_fn` is applied to the
aggregate data after megacounty aggregation, and can perform any final
calculations necessary.}

\item{cw_list}{Named list of geographic crosswalks, each of which maps a zip5
to a geographic level such as county or state. Aggregates will be produced
for each geographic level.}
}
\value{
list of data frame of individual response data and user-set data
frame of desired aggregations

list of data frame of individual response data and user-set data
frame of desired aggregations
}
\description{
Post-process aggregations and data to make more generic

Post-process aggregations and data to make more generic
}
