% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_run.R
\name{run_contingency_tables}
\alias{run_contingency_tables}
\title{Run the contingency table production pipeline}
\usage{
run_contingency_tables(params)
}
\arguments{
\item{params}{Params object produced by read_params}
}
\value{
none
}
\description{
Run the contingency table production pipeline
}
