% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_run.R
\name{set_aggs}
\alias{set_aggs}
\title{Make tables specifying aggregations to output}
\usage{
set_aggs()
}
\value{
named list
}
\description{
Each row represents one aggregate to report. `name` is the aggregate's base
column name.`metric` is the column of `df` containing the response value.
`group_by` is a list of variables used to perform the aggregations over.
`compute_fn` is the function that computes the aggregate response given many
rows of data. `post_fn` is applied to the aggregate data and can perform any
final calculations necessary.
}
\details{
Listing no groupby vars implicitly computes aggregations at the national
level only. Any multiple choice and multi-select questions used in the
aggregations should be recoded to be descriptive in
`contingency_variables::reformat_responses`.

Compute functions must be one of the `compute_*` set (or another function
with similar format can be created). Post-processing functions should be one
of the `jeffreys_*` set or post_convert_count_to_pct or the identity `I`,
which does not modify the data.
}
