% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{start_of_prev_full_month}
\alias{start_of_prev_full_month}
\title{Get the date of the first day of the previous month}
\usage{
start_of_prev_full_month(date)
}
\arguments{
\item{date}{Date of interest}
}
\value{
Date
}
\description{
Get the date of the first day of the previous month
}
