% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{start_of_prev_full_week}
\alias{start_of_prev_full_week}
\title{Get the date of the first day of the previous epiweek}
\usage{
start_of_prev_full_week(date)
}
\arguments{
\item{date}{Date of interest}
}
\value{
Date
}
\description{
Get the date of the first day of the previous epiweek
}
