% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_aggregate.R
\name{summarize_aggregations_group}
\alias{summarize_aggregations_group}
\title{Produce estimates for all indicators in a specific target group.}
\usage{
summarize_aggregations_group(group_df, aggregations, params)
}
\arguments{
\item{group_df}{Data frame containing weight fields and all metric fields.}

\item{aggregations}{Aggregations to report. See `produce_aggregates()`.}

\item{params}{Named list of configuration options.}
}
\value{
named list where each element is the val, se, n, n effective, or
  represented population for a given aggregation
}
\description{
Produce estimates for all indicators in a specific target group.
}
