% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{summarize_hh_count}
\alias{summarize_hh_count}
\title{Summarize CLI and ILI household variables at a geographic level.}
\usage{
summarize_hh_count(
  df,
  crosswalk_data,
  indicators,
  geo_level,
  days,
  params,
  smooth_days = 0L
)
}
\arguments{
\item{df}{a data frame of survey responses}

\item{crosswalk_data}{An aggregation, such as zip => county or zip => state,
as a data frame with a "zip5" column to join against.}

\item{indicators}{list of lists. Each constituent list has entries `metric`,
`var_weight`. Its name is the name of the indicator to report in the API.
`var_weight` is the name of the column of `df` to use for weights; `metric`
is the name of the metric to extract from the data frame.}

\item{geo_level}{the aggregation level, such as county or state, being used}

\item{days}{a vector of Dates for which we should generate response estimates}

\item{params}{a named list with entries "s_weight", "s_mix_coef",
"num_filter"}

\item{smooth_days}{integer; how many does in the past should be pooled into
the estimate of a day}
}
\description{
See `summarize_binary()` for a description of the architecture here and its
tradeoffs made in favor of high speed.
}
