% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{update_archive}
\alias{update_archive}
\title{Save new data to the archive}
\usage{
update_archive(df, archive, params)
}
\arguments{
\item{df}{a data frame with columns "token", "start_dr" and "ResponseID"}

\item{archive}{a data frame previously loaded by \code{load_archive}}

\item{params}{a named list, read from read_params. Must contain an element
named archive_dir (containing the file "data_agg.Rds").}
}
\description{
The archive contains:
}
\details{
- all previously seen tokens

- all prior rows of input data (before filtering for aggregation or sharing),
  except for rows with duplicate tokens.
}
