% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_dev.R
\name{update_params}
\alias{update_params}
\title{Update date and input files settings from params file}
\usage{
update_params(params)
}
\arguments{
\item{params}{Params object produced by read_params}
}
\value{
a named list of parameters values
}
\description{
Update date and input files settings from params file
}
