% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{write_cid}
\alias{write_cid}
\title{Write a file containing the new tokens seen in the dataset}
\usage{
write_cid(data, type_name, params)
}
\arguments{
\item{data}{a data frame containing a column named "token"}

\item{type_name}{character value used for naming the output file}

\item{params}{a named listed; must contain entries "start_time", "end_time", and
"weights_out_dir". These are used in constructing the path where the
output data will be stored.}
}
\description{
Write a file containing the new tokens seen in the dataset
}
