% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_write.R
\name{write_contingency_tables}
\alias{write_contingency_tables}
\title{Write csv file for sharing with researchers.}
\usage{
write_contingency_tables(data, params, geo_type, groupby_vars)
}
\arguments{
\item{data}{a data frame to save; must contain the columns in
`groupby_vars`.}

\item{params}{a named list, containing the values:
"export_dir" - directory where the csv should be saved
"static_dir" - directory where the state lookup file is
"aggregate_range" - "month", "week", etc.
"start_date" - start date of the aggregate range
"end_date" - end date of the aggregate range}

\item{geo_type}{name of the geographic level; used for naming the output file}

\item{groupby_vars}{character vector of column names used for grouping to
calculate aggregations; used for naming the output file}
}
\description{
CSV name includes date specifying start of time period aggregated, geo level,
and grouping variables.
}
